Move i_watch support code from worker.c to inotify-watch.c
